Split the native renderer into shim table, op-family tiers and finish seam
