+++ crates/almide-mir/src/render_native.rs
@@ -28,7 +28,7 @@ use std::fmt::Write as _;
 
 /// The native type a MIR value renders to.
 #[derive(Clone, Copy, PartialEq, Eq, Debug)]
-enum NTy {
+pub(crate) enum NTy {
     I64,
     /// An OWNED `String` local (fresh from a literal alloc, a heap-returning
     /// call, or a clone).
@@ -75,11 +75,11 @@ pub enum NativeSigKind {
 /// rejected anything outside these kinds).
 pub type NativeSigs = std::collections::BTreeMap<String, (Vec<NativeSigKind>, Option<NativeSigKind>)>;
 
-fn wall(msg: impl Into<String>) -> LowerError {
+pub(crate) fn wall(msg: impl Into<String>) -> LowerError {
     LowerError::Unsupported(msg.into())
 }
 
-fn var(v: ValueId) -> String {
+pub(crate) fn var(v: ValueId) -> String {
     format!("v{}", v.0)
 }
 
@@ -89,7 +89,7 @@ const IDX_GET_SHIM: &str = "fn almide_idx_get(v: &[i64], i: i64) -> i64 {\n     
 const IDX_SET_SHIM: &str = "fn almide_idx_set(v: &mut Vec<i64>, i: i64, x: i64) {\n        if i < 0 || i as usize >= v.len() { eprintln!(\"Error: index out of bounds\"); std::process::exit(1); }\n        v[i as usize] = x;\n}";
 
 /// Borrow a stringy value as `&str` for a call argument.
-fn as_str_arg(code: &str, t: NTy) -> String {
+pub(crate) fn as_str_arg(code: &str, t: NTy) -> String {
     match t {
         NTy::Str => format!("&{code}"),
         NTy::StrRef => code.to_string(),
@@ -102,7 +102,7 @@ fn as_str_arg(code: &str, t: NTy) -> String {
 /// Read a MIR scalar as a real `f64`: an I64 local holds the f64 BITS (the MIR
 /// Float convention — every float literal is a `ConstInt` of the bits), an F64
 /// local IS the value. Bit-exact either way.
-fn as_f64_arg(code: &str, t: NTy) -> Result<String, LowerError> {
+pub(crate) fn as_f64_arg(code: &str, t: NTy) -> Result<String, LowerError> {
     match t {
         NTy::F64 => Ok(code.to_string()),
         NTy::I64 => Ok(format!("f64::from_bits({code} as u64)")),
@@ -115,132 +115,8 @@ fn as_f64_arg(code: &str, t: NTy) -> Result<String, LowerError> {
 /// is adding to the trusted floor — keep it tiny; everything else walls. Every
 /// addition needs a differential-corpus row in the same PR
 /// (tests/native_v1_differential_test.rs).
-fn shim(name: &str) -> Option<(&'static [NTy], Option<NTy>, &'static str)> {
-    match name {
-        "int.to_string" => Some((
-            &[NTy::I64],
-            Some(NTy::Str),
-            "fn rt_int_to_string(n: i64) -> String { n.to_string() }",
-        )),
-        "print_str" => Some((
-            &[NTy::Str],
-            None,
-            "fn rt_print_str(s: &str) { println!(\"{}\", s); }",
-        )),
-        // The §13 abort convention's message channel (assert desugar, time-ctor
-        // negative trap): stderr line, exact v0 oracle behavior.
-        "eprintln" => Some((
-            &[NTy::Str],
-            None,
-            "fn rt_eprintln(s: &str) { eprintln!(\"{}\", s); }",
-        )),
-        "__str_concat" => Some((
-            &[NTy::Str, NTy::Str],
-            Some(NTy::Str),
-            "fn rt_str_concat(a: &str, b: &str) -> String { [a, b].concat() }",
-        )),
-        "string.eq" => Some((
-            &[NTy::Str, NTy::Str],
-            Some(NTy::I64),
-            "fn rt_string_eq(a: &str, b: &str) -> i64 { (a == b) as i64 }",
-        )),
-        "string.len" => Some((
-            // Codepoint count, NOT byte length (C-016 discipline).
-            &[NTy::Str],
-            Some(NTy::I64),
-            "fn rt_string_len(s: &str) -> i64 { s.chars().count() as i64 }",
-        )),
-        // String predicates/transforms: each shim is the EXACT v0 native oracle
-        // expression (runtime/rs/src/string.rs delegates to Rust std the same way),
-        // so the differential gate pins byte-equality, and C-016/C-019/C-020's
-        // full-Unicode discipline carries over unchanged.
-        "string.contains" => Some((
-            &[NTy::Str, NTy::Str],
-            Some(NTy::I64),
-            "fn rt_string_contains(s: &str, sub: &str) -> i64 { s.contains(sub) as i64 }",
-        )),
-        "string.starts_with" => Some((
-            &[NTy::Str, NTy::Str],
-            Some(NTy::I64),
-            "fn rt_string_starts_with(s: &str, p: &str) -> i64 { s.starts_with(p) as i64 }",
-        )),
-        "string.ends_with" => Some((
-            &[NTy::Str, NTy::Str],
-            Some(NTy::I64),
-            "fn rt_string_ends_with(s: &str, p: &str) -> i64 { s.ends_with(p) as i64 }",
-        )),
-        "string.to_upper" => Some((
-            &[NTy::Str],
-            Some(NTy::Str),
-            "fn rt_string_to_upper(s: &str) -> String { s.to_uppercase() }",
-        )),
-        "string.to_lower" => Some((
-            &[NTy::Str],
-            Some(NTy::Str),
-            "fn rt_string_to_lower(s: &str) -> String { s.to_lowercase() }",
-        )),
-        "string.trim" => Some((
-            &[NTy::Str],
-            Some(NTy::Str),
-            "fn rt_string_trim(s: &str) -> String { s.trim().to_string() }",
-        )),
-        "string.repeat" => Some((
-            &[NTy::Str, NTy::I64],
-            Some(NTy::Str),
-            // The SAME clamp + ceiling as the v0 runtime and the wasm self-host
-            // (stdlib/string_repeat.almd): a negative count is the empty string
-            // (C-054), and a result past the shared 2^31 ceiling aborts in the
-            // T6 form. The bare `s.repeat(n as usize)` turned `repeat(s, -1)`
-            // into a capacity-overflow PANIC (exit 101) on this leg while the
-            // wasm leg printed normally — a crash-form divergence the C-161
-            // rule forbids (differential fuzz, seed 1785015406589852000 index
-            // 1012). Keep in sync with ALMIDE_REPEAT_MAX_BYTES.
-            "fn rt_string_repeat(s: &str, n: i64) -> String {\n    let n = n.max(0);\n    if (s.len() as i64).saturating_mul(n) > (1i64 << 31) {\n        eprintln!(\"Error: repeat result too large\");\n        std::process::exit(1);\n    }\n    s.repeat(n as usize)\n}",
-        )),
-        "string.cmp" => Some((
-            // Byte-wise lexicographic, -1/0/1 (C-019: rt_string_extra cmp = native oracle).
-            &[NTy::Str, NTy::Str],
-            Some(NTy::I64),
-            "fn rt_string_cmp(a: &str, b: &str) -> i64 {\n    match a.cmp(b) { std::cmp::Ordering::Less => -1, std::cmp::Ordering::Equal => 0, std::cmp::Ordering::Greater => 1 }\n}",
-        )),
-        "float.to_string" => Some((
-            // The EXACT v0 native oracle (runtime/rs/src/float.rs::almide_rt_float_to_string):
-            // shortest round-trip Display, integral values forced to a `.0` tail.
-            &[NTy::F64],
-            Some(NTy::Str),
-            "fn rt_float_to_string(n: f64) -> String {\n    let s = format!(\"{}\", n);\n    if n.fract() == 0.0 && !s.contains('.') && !s.contains(\"inf\") && !s.contains(\"NaN\") {\n        format!(\"{}.0\", s)\n    } else {\n        s\n    }\n}",
-        )),
-        "__chk_div" => Some((
-            &[NTy::I64, NTy::I64],
-            Some(NTy::I64),
-            "fn rt_chk_div(a: i64, b: i64) -> i64 {\n    if b == 0 { eprintln!(\"Error: division by zero\"); std::process::exit(1); }\n    if a == i64::MIN && b == -1 { eprintln!(\"Error: integer overflow\"); std::process::exit(1); }\n    a / b\n}",
-        )),
-        "__chk_mod" => Some((
-            &[NTy::I64, NTy::I64],
-            Some(NTy::I64),
-            // v0's `almide_mod` macro prints "division by zero" for a zero rhs (mod and
-            // div share the message — the C-002 oracle text); keep byte parity.
-            "fn rt_chk_mod(a: i64, b: i64) -> i64 {\n    if b == 0 { eprintln!(\"Error: division by zero\"); std::process::exit(1); }\n    if a == i64::MIN && b == -1 { eprintln!(\"Error: integer overflow\"); std::process::exit(1); }\n    a % b\n}",
-        )),
-        // The unsigned 64-bit lane (#872): the i64 slot carries the u64 bit
-        // pattern. Same divide-by-zero message bytes; no MIN÷-1 case unsigned.
-        "__chk_div_u" => Some((
-            &[NTy::I64, NTy::I64],
-            Some(NTy::I64),
-            "fn rt_chk_div_u(a: i64, b: i64) -> i64 {\n    if b == 0 { eprintln!(\"Error: division by zero\"); std::process::exit(1); }\n    ((a as u64) / (b as u64)) as i64\n}",
-        )),
-        "__chk_mod_u" => Some((
-            &[NTy::I64, NTy::I64],
-            Some(NTy::I64),
-            "fn rt_chk_mod_u(a: i64, b: i64) -> i64 {\n    if b == 0 { eprintln!(\"Error: division by zero\"); std::process::exit(1); }\n    ((a as u64) % (b as u64)) as i64\n}",
-        )),
-        _ => None,
-    }
-}
-
-fn shim_rust_name(name: &str) -> String {
-    format!("rt_{}", name.trim_start_matches("__").replace('.', "_"))
-}
+use crate::render_native_op_families::{render_native_meter_op, render_native_result_op, render_native_termination_op, render_native_scalar_op, render_native_float_op};
+use crate::render_native_shims::{shim, shim_rust_name, CUT_RET_MARKER, FUEL_LT0_SHIM, CHARGE_DYN_SHIM, TIMEOUT_SHIM, COUNTER_SHIM, CHARGE_SHIM, BUDGET_SHIM};
 
 /// Render a whole MIR program to a self-contained Rust source, or WALL.
 pub fn try_render_native_program(prog: &MirProgram, sigs: &NativeSigs) -> Result<String, LowerError> {
@@ -417,153 +293,17 @@ fn render_fn(
                     },
                 )?
             }
-            Op::Charge { site, cost } => {
-                used_shims.push(COUNTER_SHIM);
-                used_shims.push(CHARGE_SHIM);
-                let tr = crate::charge_probe::probe_enabled();
-                used_shims.push(FUEL_LT0_SHIM);
-                line!("__almd_charge({site}, {cost}, {tr});");
-                // T1-1 strict cut (see the wasm arm): the dummy return value's
-                // type is known only after the body typed `func.ret`, so a
-                // marker is emitted here and patched at the end of render_fn.
-                line!("if __almd_fuel_lt0() {{ {CUT_RET_MARKER} }}");
-                // T5-1: the wall-deadline check rides the same cut mechanism.
-                if crate::charge_probe::timeout_used() {
-                    used_shims.push(TIMEOUT_SHIM.as_str());
-                    line!("if __almd_wall_hit() {{ {CUT_RET_MARKER} }}");
-                }
-            }
-            // T3-5 dynamic charge — the native twin of the wasm arm above:
-            // 1 + byte_len/16 of the result string, same trace + cut rules.
-            Op::ChargeDyn { site, src } => {
-                used_shims.push(COUNTER_SHIM);
-                used_shims.push(CHARGE_SHIM);
-                used_shims.push(CHARGE_DYN_SHIM);
-                used_shims.push(FUEL_LT0_SHIM);
-                let tr = crate::charge_probe::probe_enabled();
-                let sref = match tys.get(src) {
-                    Some(NTy::Str | NTy::StrRef) => format!("{}.len() as i64", var(*src)),
-                    other => {
-                        return Err(wall(format!(
-                            "native: ChargeDyn over a non-string value ({other:?})"
-                        )))
-                    }
-                };
-                line!("__almd_charge_dyn({site}, {sref}, {tr});");
-                line!("if __almd_fuel_lt0() {{ {CUT_RET_MARKER} }}");
-                if crate::charge_probe::timeout_used() {
-                    used_shims.push(TIMEOUT_SHIM.as_str());
-                    line!("if __almd_wall_hit() {{ {CUT_RET_MARKER} }}");
-                }
-            }
-            // The §13 termination convention's exit half (assert desugar tail,
-            // time-ctor negative trap): a user exit code, no message of its own.
-            Op::Prim { kind: crate::PrimKind::ProcExit, dst: None, args } => {
-                line!("std::process::exit({} as i32);", var(args[0]));
-            }
-            // The §13 termination convention's MESSAGE half — the err-abort
-            // window of a `?`-propagation in main (`Handle(msg)` + `Die`).
-            // The native render has no address model: a handle over a STRING
-            // local is an alias (the value IS the string), and `Die` is the
-            // wasm `$__die` twin — the string to stderr verbatim, exit 1.
-            Op::Prim { kind: crate::PrimKind::Handle, dst: Some(d), args }
-                if matches!(tys.get(&args[0]), Some(NTy::Str | NTy::StrRef)) =>
-            {
-                let src = match tys.get(&args[0]) {
-                    Some(NTy::Str) => format!("{}.as_str()", var(args[0])),
-                    _ => var(args[0]).to_string(),
-                };
-                tys.insert(*d, NTy::StrRef);
-                line!("let {}: &str = {};", var(*d), src);
-            }
-            Op::Prim { kind: crate::PrimKind::Die, dst: None, args }
-                if matches!(tys.get(&args[0]), Some(NTy::Str | NTy::StrRef)) =>
-            {
-                line!("eprint!(\"{{}}\", {});", var(args[0]));
-                line!("std::process::exit(1);");
-            }
-            // ── T1-3 native Result carrier (native_result_rewrite) ──
-            Op::Prim { kind: crate::PrimKind::ResMakeOk, dst: Some(d), args } => {
-                tys.insert(*d, NTy::Res);
-                line!("let {}: Result<i64, String> = Ok({});", var(*d), var(args[0]));
-            }
-            Op::Prim { kind: crate::PrimKind::ResMakeErrStr, dst: Some(d), args } => {
-                let src = match tys.get(&args[0]) {
-                    Some(NTy::Str) => format!("{}.clone()", var(args[0])),
-                    Some(NTy::StrRef) => format!("{}.to_string()", var(args[0])),
-                    other => {
-                        return Err(wall(format!(
-                            "native: ResMakeErrStr over a non-string payload ({other:?})"
-                        )))
-                    }
-                };
-                tys.insert(*d, NTy::Res);
-                line!("let {}: Result<i64, String> = Err({});", var(*d), src);
-            }
-            Op::Prim { kind: crate::PrimKind::ResTag, dst: Some(d), args } => {
-                tys.insert(*d, NTy::I64);
-                line!("let {}: i64 = {}.is_err() as i64;", var(*d), var(args[0]));
-            }
-            Op::Prim { kind: crate::PrimKind::ResOkScalar, dst: Some(d), args } => {
-                tys.insert(*d, NTy::I64);
-                line!(
-                    "let {}: i64 = match &{} {{ Ok(x) => *x, Err(_) => 0 }};",
-                    var(*d),
-                    var(args[0])
-                );
-            }
-            Op::Prim { kind: crate::PrimKind::ResErrStr, dst: Some(d), args } => {
-                // A BORROW of the Err payload (the verifier aliases it to the
-                // Result's object); "" on the Ok side (unreached — the tag
-                // dispatch guards it).
-                tys.insert(*d, NTy::StrRef);
-                line!(
-                    "let {}: &str = match &{} {{ Err(e) => e.as_str(), Ok(_) => \"\" }};",
-                    var(*d),
-                    var(args[0])
-                );
-            }
-            Op::Prim { kind: crate::PrimKind::TimeoutEnter, dst: Some(d), args } => {
-                used_shims.push(TIMEOUT_SHIM.as_str());
-                tys.insert(*d, NTy::I64);
-                line!("let {} = __almd_timeout_enter({});", var(*d), var(args[0]));
-            }
-            Op::Prim { kind: crate::PrimKind::TimeoutExit, dst: Some(d), args } => {
-                used_shims.push(TIMEOUT_SHIM.as_str());
-                tys.insert(*d, NTy::I64);
-                line!("let {} = __almd_timeout_exit({});", var(*d), var(args[0]));
-            }
-            Op::Prim { kind: crate::PrimKind::TimeoutHit, dst: Some(d), .. } => {
-                used_shims.push(TIMEOUT_SHIM.as_str());
-                tys.insert(*d, NTy::I64);
-                line!("let {} = __almd_timeout_hit();", var(*d));
-            }
-            Op::Prim { kind: crate::PrimKind::BudgetEnter, dst: Some(d), args } => {
-                used_shims.push(COUNTER_SHIM);
-                used_shims.push(BUDGET_SHIM.as_str());
-                tys.insert(*d, NTy::I64);
-                line!("let {} = __almd_budget_enter({});", var(*d), var(args[0]));
-            }
-            Op::Prim { kind: crate::PrimKind::BudgetExhausted, dst: Some(d), .. } => {
-                used_shims.push(COUNTER_SHIM);
-                used_shims.push(BUDGET_SHIM.as_str());
-                tys.insert(*d, NTy::I64);
-                line!("let {} = __almd_budget_exhausted();", var(*d));
-            }
-            Op::Prim { kind: crate::PrimKind::BudgetExit, dst: Some(d), args } => {
-                used_shims.push(COUNTER_SHIM);
-                used_shims.push(BUDGET_SHIM.as_str());
-                tys.insert(*d, NTy::I64);
-                line!("let {} = __almd_budget_exit({});", var(*d), var(args[0]));
-            }
-            Op::Prim { kind: crate::PrimKind::BudgetSpend, dst: Some(d), .. } => {
-                used_shims.push(COUNTER_SHIM);
-                used_shims.push(BUDGET_SHIM.as_str());
-                tys.insert(*d, NTy::I64);
-                line!("let {} = __almd_budget_spend();", var(*d));
-            }
             other => {
                 let handled = render_native_call_op(
+                    other,
+                    crate::render_native::OpSink { tys: &mut tys, out: &mut out, indent, used_shims },
+                )? || render_native_meter_op(
+                    other,
+                    crate::render_native::OpSink { tys: &mut tys, out: &mut out, indent, used_shims },
+                )? || render_native_result_op(
+                    other,
+                    crate::render_native::OpSink { tys: &mut tys, out: &mut out, indent, used_shims },
+                )? || render_native_termination_op(
                     other,
                     crate::render_native::OpSink { tys: &mut tys, out: &mut out, indent, used_shims },
                 )? || render_native_scalar_op(
@@ -574,7 +314,7 @@ fn render_fn(
                     let detail = if let Op::Prim { kind, .. } = other {
                         format!("Prim {kind:?}")
                     } else {
-                        op_name(other).to_string()
+                        op_name(other)
                     };
                     return Err(wall(format!(
                         "native: op {detail:?} in `{}` — outside the rung subset",
@@ -588,6 +328,20 @@ fn render_fn(
     if !if_stack.is_empty() {
         return Err(wall("native: unbalanced IfThen/EndIf markers"));
     }
+    finish_native_fn(func, sigs, is_main, &mut tys, out)
+}
+
+/// The tail of [`render_fn`]: the lifted-effect `Ok(..)` wrap decision, the
+/// signature (renderable only after the body typed `func.ret`), the trailing
+/// return expression, and the T1-1 strict-cut marker patch (the typed
+/// default is known only here).
+fn finish_native_fn(
+    func: &MirFunction,
+    sigs: &NativeSigs,
+    is_main: bool,
+    tys: &mut BTreeMap<ValueId, NTy>,
+    mut out: String,
+) -> Result<(String, Option<NTy>), LowerError> {
 
     // A LIFTED effect fn (declared scalar ret, wrapped carrier ABI — the sigs
     // table widening in the pipeline): the body computes the raw scalar and
@@ -633,6 +387,7 @@ fn render_fn(
     }
     Ok((format!("{sig}{out}"), ret_nty))
 }
+
 
 /// The trailing return expression: a borrowed param is moved out as a fresh
 /// owned value; everything else returns the local directly.
@@ -750,88 +505,6 @@ fn render_native_call_op(op: &Op, s: OpSink<'_>) -> Result<bool, LowerError> {
     Ok(true)
 }
 
-/// One VALUE-PRODUCING op (const, alloc, dup, list literal, int/float
-/// arithmetic, local rebind) rendered into the sink. `Ok(false)` = not this
-/// tier's op — the caller tries the flow tier, then walls. Arm bodies are
-/// verbatim from the former inline [`render_fn`] op match.
-fn render_native_scalar_op(op: &Op, s: OpSink<'_>) -> Result<bool, LowerError> {
-    let OpSink { tys, out, indent, used_shims } = s;
-    macro_rules! line {
-        ($($arg:tt)*) => {{
-            for _ in 0..indent { out.push_str("    "); }
-            writeln!(out, $($arg)*).unwrap();
-        }};
-    }
-    match op {
-        Op::ConstInt { dst, value } => {
-            tys.insert(*dst, NTy::I64);
-            line!("let mut {}: i64 = {}i64;", var(*dst), value);
-        }
-        Op::Alloc { dst, init, .. } => match init {
-            Init::Str(s) => {
-                tys.insert(*dst, NTy::Str);
-                line!("let mut {}: String = String::from({s:?});", var(*dst));
-            }
-            other => return Err(wall(format!("native: Alloc {other:?} — outside the rung subset"))),
-        },
-        Op::Dup { dst, src } => render_dup(dst, src, tys, out, indent)?,
-        // Rung-4 scalar-list literal: the natural Vec spelling. Elements are raw
-        // i64 slot values (the wasm leg stores the same bits).
-        Op::ListLit { dst, elems } => {
-            for e in elems {
-                if tys.get(e) != Some(&NTy::I64) {
-                    return Err(wall("native: ListLit with a non-scalar element"));
-                }
-            }
-            tys.insert(*dst, NTy::Vec);
-            let items = elems.iter().map(|e| var(*e)).collect::<Vec<_>>().join(", ");
-            line!("let mut {}: Vec<i64> = vec![{items}];", var(*dst));
-        }
-        Op::SetLocal { local, src } => render_set_local(local, src, tys, out, indent)?,
-        Op::IntBinOp { dst, op, a, b } => {
-            tys.insert(*dst, NTy::I64);
-            let rendered = render_int_binop(op, *a, *b, used_shims)?;
-            line!("let mut {}: i64 = {};", var(*dst), rendered);
-        }
-        _ => return render_native_float_op(op, OpSink { tys, out, indent, used_shims }),
-    }
-    Ok(true)
-}
-
-/// Rung-5 float floor: MIR floats are i64 BITS; native computes in real
-/// f64. Every op below is IEEE-754-exact on both targets (hardware ops,
-/// identical bit results), so byte-identity holds through
-/// `float.to_string`. Min/Max/CopySign are excluded: Rust's `f64::min`
-/// NaN semantics differ from wasm `f64.min` (they only occur inside
-/// self-host bodies, which never render natively).
-fn render_native_float_op(op: &Op, s: OpSink<'_>) -> Result<bool, LowerError> {
-    let OpSink { tys, out, indent, .. } = s;
-    macro_rules! line {
-        ($($arg:tt)*) => {{
-            for _ in 0..indent { out.push_str("    "); }
-            writeln!(out, $($arg)*).unwrap();
-        }};
-    }
-    match op {
-        Op::Prim { kind: crate::PrimKind::FloatBin(op), dst: Some(d), args } if args.len() == 2 => {
-            render_float_bin(op, d, args, tys, out, indent)?
-        }
-        // `float.from_int` — int (i64) to f64, carried per the float floor.
-        Op::Prim { kind: crate::PrimKind::F64FromInt, dst: Some(d), args } if args.len() == 1 => {
-            tys.insert(*d, NTy::F64);
-            line!("let mut {}: f64 = ({} as f64);", var(*d), var(args[0]));
-        }
-        Op::Prim { kind: crate::PrimKind::FloatUn(op), dst: Some(d), args } if args.len() == 1 => {
-            render_float_un(op, d, args, tys, out, indent)?
-        }
-        Op::Prim { kind: crate::PrimKind::FloatCmp(op), dst: Some(d), args } if args.len() == 2 => {
-            render_float_cmp(op, d, args, tys, out, indent)?
-        }
-        _ => return Ok(false),
-    }
-    Ok(true)
-}
-
 /// One CONTROL-FLOW / ownership op (if-as-value markers, loops, drops, the
 /// bookkeeping no-ops) rendered into the buffer. `Ok(false)` = not this tier's
 /// op — the caller walls. Arm bodies are verbatim from the former inline
@@ -941,7 +614,7 @@ fn render_native_drop_op(
 
 /// `Op::Dup` — mint a fresh handle from `src` per its NTy (verbatim arm body
 /// extracted from [`render_fn`]; see that function for the op-loop context).
-fn render_dup(
+pub(crate) fn render_dup(
     dst: &ValueId,
     src: &ValueId,
     tys: &mut BTreeMap<ValueId, NTy>,
@@ -993,146 +666,3 @@ fn render_dup(
 include!("render_native_b.rs");
 
 
-/// T1-1: the strict-cut return marker — every Charge site emits
-/// `if __almd_fuel_lt0() { <marker> }` and `render_fn` patches the marker
-/// with a `return <default of the fn's ret type>;` once the ret NTy is known
-/// (the same late-patch technique as the if-value JOIN markers).
-const CUT_RET_MARKER: &str = "/*__CUT_RET__*/";
-
-/// The exhaustion read the strict cut branches on.
-const FUEL_LT0_SHIM: &str =
-    "fn __almd_fuel_lt0() -> bool { __ALMD_FUEL.with(|f| f.get()) < 0 }";
-
-/// T3-5: the dynamic (size-proportional) charge — 1 + len/16, same trace
-/// arithmetic as the static charge.
-const CHARGE_DYN_SHIM: &str = "fn __almd_charge_dyn(site: i64, len: i64, trace: bool) {
-    __ALMD_FUEL.with(|f| f.set(f.get().wrapping_sub(1 + (len >> 4))));
-    if trace {
-        __ALMD_TRACE.with(|t| t.set(t.get().wrapping_mul(1000003).wrapping_add(site)));
-    }
-}";
-
-/// T5-1: the wall-deadline shims (fan.timeout). The clock is monotonic
-/// (`Instant` since first use); in REPLAY mode (`ALMIDE_OMEGA` baked at
-/// compile time) the clock is never read — the baked ordinal decides the
-/// cut. RECORD mode (`ALMIDE_OMEGA_RECORD=1` baked) prints
-/// `__ALMD_OMEGA <ord>` on stderr at each fired region exit.
-static TIMEOUT_SHIM: std::sync::LazyLock<String> = std::sync::LazyLock::new(|| {
-    TIMEOUT_SHIM_TEMPLATE
-        .replace("__ALMD_OMEGA_V__", &crate::charge_probe::omega_replay().to_string())
-        .replace("__ALMD_OMEGA_REC__", if crate::charge_probe::omega_record() { "true" } else { "false" })
-});
-
-const TIMEOUT_SHIM_TEMPLATE: &str = "const __ALMD_OMEGA: i64 = __ALMD_OMEGA_V__;
-const __ALMD_OMEGA_RECORD: bool = __ALMD_OMEGA_REC__;
-static __ALMD_T0: std::sync::OnceLock<std::time::Instant> = std::sync::OnceLock::new(); // wasm-safe: GENERATED native-only shim (fan.timeout reads the wall clock at program runtime — oracle-tier semantics, never the compile path)
-fn __almd_now_ns() -> i64 {
-    __ALMD_T0.get_or_init(std::time::Instant::now).elapsed().as_nanos() as i64 // wasm-safe: generated shim (see above)
-}
-thread_local! {
-    static __ALMD_T_DEADLINE: std::cell::Cell<i64> = const { std::cell::Cell::new(i64::MAX) };
-    static __ALMD_T_HIT: std::cell::Cell<bool> = const { std::cell::Cell::new(false) };
-    static __ALMD_T_VERDICT: std::cell::Cell<i64> = const { std::cell::Cell::new(0) };
-    static __ALMD_T_ORD: std::cell::Cell<i64> = const { std::cell::Cell::new(0) };
-}
-fn __almd_timeout_enter(d_ns: i64) -> i64 {
-    let saved = __ALMD_T_DEADLINE.with(|d| d.get());
-    let now = if __ALMD_OMEGA >= 0 { 0 } else { __almd_now_ns() };
-    let dl = now.saturating_add(d_ns);
-    if dl < saved {
-        __ALMD_T_DEADLINE.with(|d| d.set(dl));
-    }
-    saved
-}
-fn __almd_timeout_exit(saved: i64) -> i64 {
-    let hit = __ALMD_T_HIT.with(|h| h.get());
-    __ALMD_T_VERDICT.with(|v| v.set(hit as i64));
-    if __ALMD_OMEGA_RECORD && hit {
-        eprintln!(\"__ALMD_OMEGA {}\", __ALMD_T_ORD.with(|o| o.get()));
-    }
-    __ALMD_T_HIT.with(|h| h.set(false));
-    __ALMD_T_DEADLINE.with(|d| d.set(saved));
-    0
-}
-fn __almd_timeout_hit() -> i64 {
-    __ALMD_T_VERDICT.with(|v| v.get())
-}
-fn __almd_wall_hit() -> bool {
-    if __ALMD_T_DEADLINE.with(|d| d.get()) == i64::MAX {
-        return false;
-    }
-    if __ALMD_T_HIT.with(|h| h.get()) {
-        return true;
-    }
-    let ord = __ALMD_T_ORD.with(|o| { o.set(o.get() + 1); o.get() });
-    if __ALMD_OMEGA >= 0 {
-        if ord >= __ALMD_OMEGA {
-            __ALMD_T_HIT.with(|h| h.set(true));
-        }
-        return __ALMD_T_HIT.with(|h| h.get());
-    }
-    if __almd_now_ns() >= __ALMD_T_DEADLINE.with(|d| d.get()) {
-        __ALMD_T_HIT.with(|h| h.set(true));
-    }
-    __ALMD_T_HIT.with(|h| h.get())
-}";
-
-/// Stage 1 probe shim: fuel/trace thread-locals + the charge fn + the guard
-/// that prints the triple's (consumed, trace) legs on main exit. Same hash
-/// arithmetic as the wasm leg (wrapping i64, trace*1000003+site).
-const COUNTER_SHIM: &str = "thread_local! {
-    static __ALMD_FUEL: std::cell::Cell<i64> = const { std::cell::Cell::new(i64::MAX) };
-    static __ALMD_FUEL_ENTRY: std::cell::Cell<i64> = const { std::cell::Cell::new(0) };
-    static __ALMD_B_VERDICT: std::cell::Cell<i64> = const { std::cell::Cell::new(0) };
-    static __ALMD_B_SPEND: std::cell::Cell<i64> = const { std::cell::Cell::new(0) };
-    static __ALMD_TRACE: std::cell::Cell<i64> = const { std::cell::Cell::new(0) };
-}";
-
-/// The probe charge fn: fuel counts DOWN (consumed = MAX - fuel); the trace is
-/// the order-sensitive hash. Only probe builds call it with tracing on.
-const CHARGE_SHIM: &str = "fn __almd_charge(site: i64, cost: i64, trace: bool) {
-    __ALMD_FUEL.with(|f| f.set(f.get().wrapping_sub(cost)));
-    if trace {
-        __ALMD_TRACE.with(|t| t.set(t.get().wrapping_mul(1000003).wrapping_add(site)));
-    }
-}
-struct __AlmdProbeGuard;
-impl Drop for __AlmdProbeGuard {
-    fn drop(&mut self) {
-        eprintln!(\"__ALMD_PROBE {} {}\",
-            (i64::MAX.wrapping_sub(__ALMD_FUEL.with(|f| f.get()))) as u64,
-            __ALMD_TRACE.with(|t| t.get()) as u64);
-    }
-}";
-
-/// Stage 2 budget fns — the exact wasm-leg arithmetic (min-cap, lazy verdict,
-/// streaming exit). The ns→unit divisor is injected from the single CM-1
-/// definition ([`crate::charge_probe::CM1_NS_PER_CHARGE`]) so this shim cannot
-/// drift from the wasm BudgetEnter render.
-static BUDGET_SHIM: std::sync::LazyLock<String> = std::sync::LazyLock::new(|| {
-    BUDGET_SHIM_TEMPLATE
-        .replace("__ALMD_CM1_NS__", &crate::charge_probe::CM1_NS_PER_CHARGE.to_string())
-});
-
-const BUDGET_SHIM_TEMPLATE: &str = "fn __almd_budget_enter(budget_ns: i64) -> i64 {
-    let units = budget_ns / __ALMD_CM1_NS__;
-    __ALMD_FUEL_ENTRY.with(|e| e.set(units));
-    let saved = __ALMD_FUEL.with(|f| f.get());
-    if units < saved {
-        __ALMD_FUEL.with(|f| f.set(units));
-    }
-    saved
-}
-fn __almd_budget_exhausted() -> i64 {
-    __ALMD_B_VERDICT.with(|v| v.get())
-}
-fn __almd_budget_exit(saved: i64) -> i64 {
-    __ALMD_B_VERDICT.with(|v| v.set(i64::from(__ALMD_FUEL.with(|f| f.get()) < 0)));
-    let consumed = __ALMD_FUEL_ENTRY.with(|e| e.get()) - __ALMD_FUEL.with(|f| f.get());
-    __ALMD_B_SPEND.with(|s| s.set(consumed));
-    __ALMD_FUEL.with(|f| f.set(saved - consumed));
-    0
-}
-fn __almd_budget_spend() -> i64 {
-    __ALMD_B_SPEND.with(|s| s.get())
-}";
